backend/requirements.txt
Django>=2.2.3,<2.3.0
djangorestframework>=3.9.4,<3.10.0
psycopg2>=2.8.3,<2.9.0
flake8>=3.7.7,<3.8.0
drf-nested-routers
django-cors-headers #integration with react



